requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.3
djangorestframework==3.15.2
idna==3.10
psycopg2==2.9.6
requests==2.32.3
slack_sdk==3.33.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
celery
django-celery-beat